Staging

+++ Assets/Scripts/Battle System/TurnManager.cs
@@ -1,13 +1,19 @@
+// TurnManager.cs
+
 using System.Collections.Generic;
-using Unity.VisualScripting;
 using UnityEngine;
 using UnityEngine.InputSystem;
 using UnityEngine.UIElements;
 
+// Vitally important class that helps facilitate a turn order for the BattleManager.
+// TODO: Instead of using GameObjects for turnOrder, I should create a helper TurnObject class, that may have entities but also AoEs and other
+// possible turns. For example, some AoEs may operate independently on a turn-by-turn basis (helps for puzzle-like mechanics)
 public class TurnManager
 {
-    List<GameObject> turnOrder = new List<GameObject>();
-    Dictionary<string, GameObject> deadUnits = new Dictionary<string, GameObject>();
+    List<GameObject> turnOrder = new List<GameObject>();    // A direct order of GameObjects (entities) in the battle, sorted by Delay. Lower Delay means they will be going before higher delays.
+    Dictionary<string, GameObject> deadUnits = new Dictionary<string, GameObject>();    // A dictionary of all dead units. Makes it easy to bring back people into the turnOrder once they are revived
+    
+    // Visual Elements. One for the turn element (top left), and the turnOrderUI parent container element
     VisualTreeAsset turnElement;
     VisualElement turnOrderUI;
 
@@ -19,24 +25,29 @@ public class TurnManager
         DisplayTurnOrder();
     }
 
+    // Grabs all players and enemies (currently just a single boss), and sorts them by a delay value.
     public void InitializeTurnOrder(List<GameObject> players, GameObject boss)
     {
         turnOrder.Clear();
 
+        // Adds all entities to a single turnOrder List
         turnOrder.AddRange(players);
         turnOrder.Add(boss);
 
+        // Calculates the initial delay for all entities
         foreach (var character in turnOrder)
         {
             character.GetComponent<EntityStats>().CalculateDelay();
         }
 
+        // Sorts them by the delay we've just calculated
         turnOrder.Sort((a, b) =>
         {
-            return a.GetComponent<EntityStats>().delay - b.GetComponent<EntityStats>().delay;
+            return (int)a.GetComponent<EntityStats>().delay - (int)b.GetComponent<EntityStats>().delay;
         });
     }
 
+    // For an entity who either gets revived, or just goes, we recalculate and insert them back
     public bool InsertUnitIntoTurn(GameObject entity)
     {
         EntityStats entityStats = entity.GetComponent<EntityStats>();
@@ -61,6 +72,7 @@ public class TurnManager
         turnOrderUI = ExcentraDatabase.TryGetDocument("battle").rootVisualElement.Q<VisualElement>("turn-order");
     }
 
+    // Clears and re-displays the appropriate turn order
     public void DisplayTurnOrder()
     {
         turnOrderUI.Clear();
@@ -70,7 +82,7 @@ public class TurnManager
             VisualElement charPortrait = charElement.Q<VisualElement>("portrait");
             Label delayLabel = charElement.Q<Label>("delay");
             charPortrait.style.backgroundImage = character.GetComponent<EntityStats>().portrait;
-            delayLabel.text = character.GetComponent<EntityStats>().delay.ToString();
+            delayLabel.text = ((int)character.GetComponent<EntityStats>().delay).ToString();
             turnOrderUI.Add(charElement);
         }
     }
@@ -99,6 +111,8 @@ public class TurnManager
         }
     }
 
+    // Calculates the delay for all entities that already exist in the turn order.
+    // When an entity goes, all other entities get their delay reduced by a percentage found in the CalculateDelay function
     public void CalculateAllDelay()
     {
         foreach (var character in turnOrder)
@@ -107,6 +121,7 @@ public class TurnManager
         }
     }
 
+    // Ends the current turn, pops the unit out of the turn order, and replaces them back in with a newly fresh calculated delay
     public void EndCurrentTurn()
     {
         GameObject currTurn = PopCurrentTurn();
